peek_word and poke_word removed in favor of peek_bytes and poke_bytes
The problem is, although both x86 and 65* architectures are little-endian, Python is big-endian. Probably because it's network order, and it makes more sense. Anyway, sometimes you'll want to get or set a value and sometimes you'll want to get or set two bytes, and both of these cases demand different handling. So just use peek_bytes and poke_bytes instead, and have the struct (or array?) module handle conversion if you need it.